SpecificationsField -> ResultsRangesField
