+++ bika/lims/browser/fields/__init__.py
@@ -29,3 +29,5 @@ from .coordinatefield import CoordinateField
 from .reflexrulefield import ReflexRuleField
 from .proxyfield import ProxyField
 from .uidreferencefield import UIDReferenceField
+from .resultrangefield import ResultRangeField
+from .resultsrangesfield import ResultsRangesField
